fix(#682): address native overlay review and terminal E2E
- Fixed F1 by moving the activity popover card and host hook behind shared
  component modules so NativeOverlayHost no longer imports agent-status feature
  code.
- Fixed F2 by making keymap setup require both PTY registration and split-view
  visibility, and by focusing terminal input through the visible session before
  typing.

Reviewers: github-codex-connector, deterministic-ci

GitHub-Review-Processed-Claude:
GitHub-Review-Superseded-Claude:
GitHub-Review-Processed-Codex-Reviews: 4676551405
GitHub-Review-Processed-Codex-Inline: 3563209536
GitHub-Review-Processed-Human-Issue:
GitHub-Review-Processed-Human-Inline:
Closes-Codex-Threads: PRRT_kwDOR06LW86QELE-
Pattern-Files-Touched:
  docs/reviews/patterns/module-boundaries.md
Pattern-Append-Decisions:
  - F1 (Keep feature code out of shared overlay host): module-boundaries.md
    (existing, theme: shared modules must not import feature-domain UI)
    rationale: NativeOverlayHost is a shared primitive renderer loaded for all
    native overlay windows, so activity UI belongs behind shared contracts.
